refactor: extract validation and metadata modules from api.rs
Move the inline validation helpers (~250 lines) and MetadataPolicy
struct (~60 lines) out of api.rs into dedicated sibling modules
(engine/validation.rs, engine/metadata.rs). api.rs re-exports both
so all downstream paths remain unchanged.

Phase 2 of the API/NAPI separation effort: api.rs drops from 1,894
to ~1,585 lines and focuses solely on the NAPI surface layer.
