+++ src/engine/api.rs
@@ -76,317 +76,15 @@ fn napi_err(env: &Env, err: LazyImageError) -> napi::Error {
     crate::error::napi_error_with_code(env, err).expect("failed to create napi error")
 }
 
+// Validation helpers are in the sibling `validation` module (src/engine/validation.rs).
+// Re-import under the `validation` namespace so all call-sites remain unchanged.
 #[cfg(feature = "napi")]
-mod validation {
-    use super::*;
-    use std::path::Path;
+use super::validation;
 
-    #[derive(Debug, PartialEq, Eq)]
-    pub enum LimitUpdate {
-        Unchanged,
-        Disable,
-        Set(u64),
-    }
-
-    fn number_label(v: f64) -> String {
-        if v.is_nan() {
-            "NaN".to_string()
-        } else if v.is_infinite() && v.is_sign_positive() {
-            "Infinity".to_string()
-        } else if v.is_infinite() && v.is_sign_negative() {
-            "-Infinity".to_string()
-        } else {
-            v.to_string()
-        }
-    }
-
-    pub fn ensure_finite_integer(
-        name: &'static str,
-        value: f64,
-    ) -> std::result::Result<i64, LazyImageError> {
-        if !value.is_finite() {
-            return Err(LazyImageError::invalid_argument(
-                name,
-                number_label(value),
-                "must be a finite number",
-            ));
-        }
-        if value.fract() != 0.0 {
-            return Err(LazyImageError::invalid_argument(
-                name,
-                value.to_string(),
-                "must be an integer",
-            ));
-        }
-        if value.abs() > (i64::MAX as f64) {
-            return Err(LazyImageError::invalid_argument(
-                name,
-                value.to_string(),
-                "is too large",
-            ));
-        }
-        Ok(value as i64)
-    }
-
-    fn sanitize_dimension(
-        name: &'static str,
-        raw: Option<f64>,
-        allow_zero: bool,
-    ) -> std::result::Result<Option<u32>, LazyImageError> {
-        match raw {
-            None => Ok(None),
-            Some(v) => {
-                let int = ensure_finite_integer(name, v)?;
-                if int < 0 {
-                    return Err(LazyImageError::invalid_resize_dimensions(Some(0), None));
-                }
-                let value_u32 = u32::try_from(int)
-                    .map_err(|_| LazyImageError::invalid_resize_dimensions(None, None))?;
-                if !allow_zero && value_u32 == 0 {
-                    return Err(LazyImageError::invalid_resize_dimensions(Some(0), None));
-                }
-                if value_u32 > crate::engine::MAX_DIMENSION {
-                    return Err(LazyImageError::invalid_resize_dimensions(
-                        Some(value_u32),
-                        None,
-                    ));
-                }
-                Ok(Some(value_u32))
-            }
-        }
-    }
-
-    pub fn sanitize_resize_dimensions(
-        width: Option<f64>,
-        height: Option<f64>,
-    ) -> std::result::Result<(Option<u32>, Option<u32>), LazyImageError> {
-        let width = sanitize_dimension("width", width, false)?;
-        let height = sanitize_dimension("height", height, false)?;
-
-        if width.is_none() && height.is_none() {
-            return Err(LazyImageError::invalid_resize_dimensions(None, None));
-        }
-
-        Ok((width, height))
-    }
-
-    pub fn sanitize_crop(
-        x: f64,
-        y: f64,
-        width: f64,
-        height: f64,
-    ) -> std::result::Result<(u32, u32, u32, u32), LazyImageError> {
-        let x_i = ensure_finite_integer("x", x)?;
-        let y_i = ensure_finite_integer("y", y)?;
-        let width_i = ensure_finite_integer("width", width)?;
-        let height_i = ensure_finite_integer("height", height)?;
-
-        if x_i < 0 || y_i < 0 {
-            return Err(LazyImageError::invalid_argument(
-                "crop offset",
-                format!("x={x_i}, y={y_i}"),
-                "must be zero or positive",
-            ));
-        }
-
-        if width_i <= 0 || height_i <= 0 {
-            let w = u32::try_from(width_i.max(0)).unwrap_or(0);
-            let h = u32::try_from(height_i.max(0)).unwrap_or(0);
-            return Err(LazyImageError::invalid_crop_dimensions(w, h));
-        }
-
-        let width_u32 = u32::try_from(width_i)
-            .map_err(|_| LazyImageError::invalid_crop_dimensions(u32::MAX, u32::MAX))?;
-        let height_u32 = u32::try_from(height_i)
-            .map_err(|_| LazyImageError::invalid_crop_dimensions(u32::MAX, u32::MAX))?;
-
-        if width_u32 > crate::engine::MAX_DIMENSION || height_u32 > crate::engine::MAX_DIMENSION {
-            return Err(LazyImageError::invalid_crop_dimensions(
-                width_u32, height_u32,
-            ));
-        }
-
-        let x_u32 = u32::try_from(x_i).map_err(|_| {
-            LazyImageError::invalid_argument("x", x_i.to_string(), "must fit into u32")
-        })?;
-        let y_u32 = u32::try_from(y_i).map_err(|_| {
-            LazyImageError::invalid_argument("y", y_i.to_string(), "must fit into u32")
-        })?;
-
-        Ok((x_u32, y_u32, width_u32, height_u32))
-    }
-
-    pub fn sanitize_quality(
-        quality: Option<f64>,
-    ) -> std::result::Result<Option<u8>, LazyImageError> {
-        match quality {
-            None => Ok(None),
-            Some(v) => {
-                let int = ensure_finite_integer("quality", v)?;
-                if !(1..=100).contains(&int) {
-                    return Err(LazyImageError::invalid_argument(
-                        "quality",
-                        int.to_string(),
-                        "must be 1-100",
-                    ));
-                }
-                let value = u8::try_from(int).map_err(|_| {
-                    LazyImageError::invalid_argument("quality", int.to_string(), "must be 1-100")
-                })?;
-                Ok(Some(value))
-            }
-        }
-    }
-
-    pub fn sanitize_concurrency(
-        concurrency: Option<f64>,
-    ) -> std::result::Result<u32, LazyImageError> {
-        match concurrency {
-            None => Ok(0),
-            Some(v) => {
-                let int = ensure_finite_integer("concurrency", v)?;
-                if int < 0 {
-                    return Err(LazyImageError::invalid_argument(
-                        "concurrency",
-                        int.to_string(),
-                        "must be >= 0",
-                    ));
-                }
-                let value = u32::try_from(int).map_err(|_| {
-                    LazyImageError::invalid_argument(
-                        "concurrency",
-                        int.to_string(),
-                        "must be within u32 range",
-                    )
-                })?;
-
-                if value > crate::engine::MAX_CONCURRENCY as u32 {
-                    return Err(LazyImageError::invalid_argument(
-                        "concurrency",
-                        value.to_string(),
-                        format!("must be 0 or 1-{}", crate::engine::MAX_CONCURRENCY),
-                    ));
-                }
-
-                Ok(value)
-            }
-        }
-    }
-
-    pub fn sanitize_limit(
-        name: &'static str,
-        value: Option<f64>,
-    ) -> std::result::Result<LimitUpdate, LazyImageError> {
-        match value {
-            None => Ok(LimitUpdate::Unchanged),
-            Some(v) => {
-                let int = ensure_finite_integer(name, v)?;
-                if int < 0 {
-                    return Err(LazyImageError::invalid_argument(
-                        name,
-                        int.to_string(),
-                        "must be >= 0",
-                    ));
-                }
-                if int == 0 {
-                    return Ok(LimitUpdate::Disable);
-                }
-                let value = u64::try_from(int).map_err(|_| {
-                    LazyImageError::invalid_argument(name, int.to_string(), "must fit within u64")
-                })?;
-                Ok(LimitUpdate::Set(value))
-            }
-        }
-    }
-
-    pub fn validate_output_path(path: &str) -> std::result::Result<(), LazyImageError> {
-        if path.trim().is_empty() {
-            return Err(LazyImageError::invalid_argument(
-                "path",
-                "<empty>",
-                "output path must not be empty",
-            ));
-        }
-
-        let path_buf = Path::new(path);
-        let parent = path_buf.parent().ok_or_else(|| {
-            LazyImageError::invalid_argument(
-                "path",
-                path.to_string(),
-                "must include a parent directory",
-            )
-        })?;
-
-        if !parent.exists() {
-            return Err(LazyImageError::invalid_argument(
-                "path",
-                path.to_string(),
-                "parent directory does not exist",
-            ));
-        }
-
-        Ok(())
-    }
-}
-
-/// Policy for which metadata to preserve in output images.
-///
-/// Centralizes the keep-ICC / keep-EXIF / strip-GPS decision so that every
-/// output path reads a single source of truth instead of resolving ad-hoc
-/// boolean combinations.
-#[derive(Clone, Debug)]
-pub struct MetadataPolicy {
-    /// Whether to preserve ICC profile in output (default: false)
-    pub icc: bool,
-    /// Whether to preserve EXIF metadata in output (default: false)
-    pub exif: bool,
-    /// Whether to keep GPS tags in EXIF (default: false — stripped for privacy)
-    pub gps: bool,
-}
-
-impl MetadataPolicy {
-    /// Default: strip all metadata for security and smaller file sizes.
-    /// GPS is stripped by default for privacy (exceeds Sharp's capabilities).
-    pub fn default_policy() -> Self {
-        Self {
-            icc: false,
-            exif: false,
-            gps: false,
-        }
-    }
-
-    /// Strip all metadata (firewall strict mode).
-    pub fn strip_all() -> Self {
-        Self {
-            icc: false,
-            exif: false,
-            gps: false,
-        }
-    }
-
-    /// Apply firewall override: if reject_metadata is true, strip everything.
-    pub fn apply_firewall(&mut self, reject_metadata: bool) {
-        if reject_metadata {
-            *self = Self::strip_all();
-        }
-    }
-
-    /// Resolve final ICC policy.
-    pub fn effective_icc(&self) -> bool {
-        self.icc
-    }
-
-    /// Resolve final EXIF policy.
-    pub fn effective_exif(&self) -> bool {
-        self.exif
-    }
-
-    /// Whether GPS tags should be stripped from EXIF.
-    /// Returns true when GPS should be stripped (i.e., gps == false).
-    pub fn strip_gps(&self) -> bool {
-        !self.gps
-    }
-}
+// MetadataPolicy lives in the sibling `metadata` module (src/engine/metadata.rs).
+// Re-export so downstream code (tasks.rs, engine.rs tests) can still reach it
+// via `super::api::MetadataPolicy` or `crate::engine::api::MetadataPolicy`.
+pub use super::metadata::MetadataPolicy;
 
 /// The main image processing engine.
 ///
